requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-socks==0.10.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
curl-cffi==0.12.0
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.2.0
fastapi==0.116.1
finvizfinance==1.1.1
frozendict==2.4.6
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
PyJWT==2.10.1
lxml==6.0.0
multidict==6.6.3
multitasking==0.0.12
numpy==2.3.1
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-socks==2.7.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.47.2
stem==1.8.2
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65